refactor: Moved stream_events to SseMessages enum method
