+++ backend/src/utils/mod.rs
@@ -6,5 +6,5 @@ pub mod jwt;
 pub mod oauth;
 pub mod openapi;
 pub mod points_counter;
-pub mod sse_handler;
+pub mod sse_message;
 pub mod success_response;
